new_coin/requirements.txt
aiohttp==3.9.3
ecdsa==0.18.0
cryptography==42.0.5
prometheus-client==0.20.0
pyyaml==6.0.1
python-dotenv==1.0.1
aiohttp-limiter==0.2.0